Progress to resolving overloads properly
Only original Fun wrappers over primitives get op_prec, and are subject to operator-like syntax.  Inlined wrappers over '+' do not get to be infix.
Overload bitshape supports a full lattice; top (dunno), 0/1 converts, or bottom (disallowed, such as Flt->Int).
Some UNRs incorrectly resolved in the past in some tests, do not now auto-resolve.  Instead they require inlining, which makes the args precise and thus resolves.
Add a prim & which has no float equivalent, and can be used in tests without needing overload resolution.

+++ src/main/java/com/cliffc/aa/node/Node.java
@@ -114,7 +114,7 @@ public abstract class Node implements Cloneable {
   public Type all_type() { return TypeErr.ALL; }
   
   // Operator precedence is only valid for ConNode of binary functions
-  public int op_prec() { return -1; }
+  public byte op_prec() { return -1; }
 
   // Hash is function+inputs, or opcode+input_uids, and is invariant over edge
   // order (so we can swap edges without rehashing)
